Moves word cracking to it's own module
