+++ test/word_cracking_test.rb
@@ -1,0 +1,18 @@
+require './test/test_helper'
+require './lib/crypter'
+require './lib/decrypter'
+
+class WordCrackingTest < Minitest::Test
+
+  def test_when_it_has_no_assumptions_it_can_still_crack_the_encryption_by_checking_for_common_words
+    skip
+    cracker = Cracker.new('keder ohulwthnw', ReliableDate.new('040895'), nil)
+    assert_equal '02715', cracker.crack
+  end
+
+  def test_method_check_all_possible_amounts_and_find_key_that_results_in_most_words
+    skip
+    cracker = Cracker.new('keder ohulwthnw', ReliableDate.new('040895'), nil)
+    assert_equal '02715', cracker.check_all_possible_amounts_and_find_key_that_results_in_most_words
+  end
+end
